pyproject.toml
[project]
name = "sermon-translation"
version = "0.1.0"
description = "Translate Christian sermon notes from English to Mandarin Chinese"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "click>=8.3.1",
    "httpx>=0.28.1",
    "openai>=2.16.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
]

[project.scripts]
sermon-translate = "sermon_translator.main:main"

[tool.setuptools.packages.find]
where = ["src"]
